requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
autopep8==2.3.1
babel==2.16.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Django==5.1.3
django-allauth==65.2.0
django-appconf==1.0.6
django-browser-reload==1.17.0
django-compressor==4.5.1
django-cotton==1.3.0
django-dotenv==1.4.2
django-filter==24.3
django-htmx==1.21.0
django-money==3.5.3
django-ninja==1.3.0
flake8==7.1.1
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
py-moneyed==3.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.0
pywatchman==2.0.0
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
setuptools==75.6.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
